Se realiza el  proceso de Comentarios ingresando pnatallas modales y mejorando los pocesos de actualizaciones en tiempo real
