+++ components/organims/Modal/ModalPublication/CommentForm/CommentForm.scss
@@ -1,0 +1,33 @@
+@import "../../../../../styles//index.scss";
+
+.comment-form {
+	position: absolute !important;
+	bottom: 0;
+	margin: 0;
+	display: flex;
+	width: 96%;
+	.field {
+		margin: 0 !important;
+		width: 100%;
+		.ui.input {
+			height: 60px;
+			input {
+				border: 0;
+				border-top: 1px solid black;
+				border-radius: 0;
+				padding-right: 80px;
+			}
+		}
+	}
+	.ui.button {
+		background-color: $primary_color;
+		color: white;
+		margin: 0;
+		position: absolute;
+		right: 0;
+		height: 100%;
+		&:hover {
+			background-color: $grey_font;
+		}
+	}
+}
