+++ components/organims/Modal/ModalPublication/Comments/Comments.scss
@@ -1,0 +1,22 @@
+@import "../../../../../styles//index.scss";
+
+.comments {
+	height: calc(100% - 95px);
+	overflow-y: scroll;
+	.comment {
+		display: flex;
+		align-items: center;
+		padding: 10px;
+		border-bottom: 1px solid $primary_color;
+		color: $grey_font;
+		img {
+			margin-right: 15px !important;
+		}
+		p {
+			margin: 0;
+			&:first-of-type {
+				font-weight: bold;
+			}
+		}
+	}
+}
