+++ components/organims/Modal/ModalPublication/ModalPublication.scss
@@ -1,0 +1,18 @@
+@import "../../../../styles//index.scss";
+
+.modal-publication {
+	height: 520px !important;
+	.ui.grid {
+		margin: 0;
+		height: 100%;
+	}
+	&__left {
+		background-position: center;
+		background-size: cover;
+		background-repeat: no-repeat;
+	}
+	&__right {
+		padding: 0 !important;
+		height: 100%;
+	}
+}
